Pruebas (1)

+++ src/edu/ujcv/progra1/RandomArray.java
@@ -5,8 +5,8 @@ import java.util.Random;
 public class RandomArray {
     Random random = new Random();
 
-    int[] randomArray(int tamano){
-        int[] arr = new int[tamano];
+    int[] randomArray(double tamano){
+        int[] arr = new int[(int) tamano];
         for (int i = 0; i < arr.length; i++) {
             arr[i] = random.nextInt();
